+++ pyrk/driver.py
@@ -17,6 +17,7 @@ from pyrk.utilities import plotter
 from pyrk.utilities.logger import pyrklog
 from pyrk.inp import sim_info
 from pyrk.utilities.ur import units
+from pyrk.utilities.h5processor import H5Processor
 import os
 
 
@@ -262,9 +263,26 @@ def main(args, curr_dir):
     log_results(si)
     out_db.close_db()
     print(si.plotdir)
-    plotter.plot(sol, si)
+
+#    plotter.plot(sol, si)
+    
+    #H5 reading
+    Sim = h5processor_setup(args)
+    Sim.plot()
+
+    
     pyrklog.critical("\nSimulation succeeded.\n")
 
+def h5processor_setup(args):
+    out     = args.plotdir
+    file    = args.outfile #the processors infile is the simulations outfile 
+    simname = [f'{os.path.splitext(os.path.basename(args.infile))[0]}']
+
+    Sim     = H5Processor(infile=file,
+                          names =simname,
+                          plotdir=out)
+    
+    return Sim
 
 """Run it as a script"""
 if __name__ == "__main__":
