+++ pyrk/utilities/tests/test_h5process.py
@@ -1,4 +1,4 @@
-from pyrk.utilities.H5Post import H5Processor
+from pyrk.utilities.h5processor import H5Processor
 import os
